Fix new bug in particle BTD (#5813)
This fixes a new issue in the back-transformed diagnostics for particles
that was introduced in #5571. The problem affected runs where not all
particle species were written into the diagnostic. In that case, the
index of the particle buffer and the index of the species in the
`MultiParticleContainer` are not the same.