Use pyamrex in spacecraft charging example (#4817)
# Overview

This PR modernizes the spacecraft charging example to use the newer
`pyamrex` interface for multifabs, instead of the fields wrapper.

The `pyamrex` interface is in principle more flexible (since it gives
access to the AMReX functions that operate on `MultiFabs` and should
have all the functionalities that the old field wrappers had. (See also
https://github.com/AMReX-Codes/pyamrex/pull/370). This is also the
interface that is documented in the [WarpX
documentation](https://warpx.readthedocs.io/en/latest/usage/workflows/python_extend.html#fields)

# Resetting the checksum

I had to reset the checksum, as they differed by `~1e-2`, which is not
negligible!

Manually looking at the fields seem to indicate that the old interface
and new interface differ in how they were handling the guard cells (at
least in the way I use them in this test). While I don't fully
understand the differences, this PR seems to produce more correct
results.
(In hindsight, it is striking that we missed the unphysical `Ez` field
when we implemented this test. The reason might be that we were focusing
`phi`, which does not exhibit this unphysical feature in such an obvious
way.)

<img width="1251" alt="Screenshot 2025-04-18 at 8 55 31 AM"
src="https://github.com/user-attachments/assets/683235aa-c03b-425c-8b55-d98016d77bdb"
/>

---------

Co-authored-by: Axel Huebl <axel.huebl@plasma.ninja>