remove erroneous use of is_tangent_to_bndy in SetRhoOrJfieldFromPEC() (#5905)
SetRhoOrJfieldFromPEC(), which is applied for both PEC and PMC
boundaries, does two things. First, the rho/J deposited to the ghost
region is reflected back into the valid cells with the appropriate sign.
Second, the guard cells are filled with values consistent with the BCs
same as the electric and magnetic fields. How to do the second step
depends solely on the is_tangent_to_bndy flag. This flag is correct for
PEC, but it is incorrect for PMC. The correct thing to use is the psign
flag.

This PR uses the psign flag for the second step instead of the
is_tangent_to_bndy flag, making this flag obsolete.