Fix checksum failure in LASY test (#5941)
In this test, the laser is polarized in `y`. 

Therefore, the `Et` field (theta direction) is large (in the default
output plane, which is the one for theta=0), while the `Er` and `Ez`
fields (radial and longitudinal directions) should be zero.

In practice, the `Er` and `Ez` fields are not quite zero due to machine
precision errors. In other tests with similar cases of almost-zero
fields, we simply do not output these fields, to avoid sporadic checksum
failures (e.g., when a compiler version is updated).

This PR does the same for the lasy test and fixes the recently-observed
failure.