Infrastructure for using full mass matrices in the implicit solvers (#5929)
This PR adds the infrastructure for using the full mass matrices in the
implicit solvers. This is a follow-on to PR #5768, where only the
diagonal terms of the diagonal matrices were computed and used. With
this new PR, the user has the option to compute the full mass matrices
which can then be used to compute the linearized current density for the
Jacobian calculation needed during the GMRES iterations of JFNK,
bypassing the need to do particle based operations.

The setup is such that the user can use the full mass matrices for
computing J during the Jacobian calculation in GMRES, and/or use the
diagonal terms for the preconditioner. How to utilize the mass matrices
is controlled by the following flags in the input deck:

`implicit_evolve.use_mass_matrices_jacobian = true`
`implicit_evolve.use_mass_matrices_pc = true`

If the second flag is true and the first flag is false, then memory is
only allocated for the mass matrix terms that are used by the PC
(diagonal terms only for now).

The ability to use the full mass matrices as described above is only
implemented for 1D and 2D Cartesian geometries and only for direct
deposit. Other geometries and other deposition schemes are deferred to
future PRs.

- [x] Merge PR #5866 to correctly set J and E/B guard cells.
- [x] Merge PR #5919 and verify that everything works for shape = 3 and
shape = 4.

A CI test will be implemented in a future PR once the full mass matrices
have been implemented for charge-conserving depositions.

---------

Co-authored-by: Justin R Angus <jangus@login39.chn.perlmutter.nersc.gov>