Fix moving-window subvolume field output (#5985)
At initialization of moving-window simulations, the diagnostic subvolume
is displaced along the direction of the moving window. This is done to
account for restarted simulations. However, the displacement is
calculated based on the trailing edge of the diagnostic subvolume rather
than the simulation volume. This causes a displacement error when the
trailing edges of the simulation volume and the diagnostic subvolume do
not coincide. This PR corrects the displacement calculation.