fields.py relies on MultiFab global indexing (#5407)
The global indexing capability for MultiFabs was moved to pyamrex. This
PR deletes the now not needed code from WarpX, relying on the capability
in pyamrex.

Note that the MultiFab wrapper is maintained since it is persistent
across load balancing, while the underlying MultiFabs are deleted and
regenerated.

This is an API breaking change since there are differences in how to
access the ghost cells when using the global indexing.

---------

Co-authored-by: Edoardo Zoni <59625522+EZoni@users.noreply.github.com>
Co-authored-by: Roelof Groenewald <40245517+roelof-groenewald@users.noreply.github.com>
Co-authored-by: Revathi  Jambunathan <41089244+RevathiJambunathan@users.noreply.github.com>