Fix energy and momentum conservation after pair wise Coulomb collisions (#5932)
Binary, pair wise, Coulomb collisions between two particles with
differing weights does not exactly conserve energy and momentum. It
would only be conserved on average over many collisions. This PR adds a
fix to the momentum of the particles that restores the energy and
momentum so that it is exactly conserved. This is done in each grid
cell, so that the energy and momentum in each cell is conserved. This
correction is done for both intra-species and inter-species collisions.
The method is described here https://doi.org/10.1016/j.jcp.2025.113927.

The implementation is designed for pairwise Coulomb collisions, but is
implemented in the binary collisions code. It does not account for the
energy and momentum of newly created particles or deleted particles.
Currently, it can only be activated when using pairwise Coulomb
collisions.

A question remains about how to handle the loss of precision that
happens during the energy correction. When the correction is small, the
calculations will take the difference of two similar large numbers,
which can lose precision. One solution is to use `long double`, but this
does not work on GPUs (where the type is not available) and Macs (where
the type is the same as double). On a Mac, in the CI test, the relative
change in energy is ~1.e-11 compared to 1.e-14 in Linux. For now, the
code uses `long double` when it is available, otherwise only `double`.

Todo:
- [x] Add documentation
- [x] Handle cases where the correction fails
- [x] Add CI test

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>