ensure time solver-dependent alpha is set correctly in the curl-curl PC. (#6022)
The curl curl PC solves the following equation:
curl ( alpha * curl ( Eg ) ) + beta * Eg = b

This PC should not be used with a non-zero alpha for time solvers that
do not treat light waves implicitly. This PR makes it such that the
alpha coefficient is set to zero for the semi-implicit and
strang-implicit time solvers.

Also, a Warning is returned if one is using this PC with time solvers
that do not treat light waves implicitly. The Warning message tells the
user to use the Jacobi PC instead for potentially better performance.