Implicit push cleanup (#6120)
This PR refactors the ImplicitPushPX() function and is a peal-off PR
from PR #5969. The purpose of this PR is to 1) reduce the size of PR
#5969 and 2) help track down something in PR #5969 that is causing small
changes with the implicit solvers even when the new suborbit model is
not being used.

I have confirmed that the changes in this PR yield identical results
with the development branch.