created an ImplicitOptions struct to pass into Evolve(). (#6116)
In future PRs, the PhysicalParticleContainer::Evolve() routine will
require more parameters specific to the implicit methods. To avoid
bloating the list of parameter passed into Evolve(), this PR creates an
ImplicitOptions structure to pass into Evolve() that owns all of the
implicit options needed by the particle push and deposit routines.

This allows direct and dynamic control of the what goes on in the
Evolve() routine from the ImplicitSolver class.