Cleanup/streamline ImplicitPushXPSubOrbits() (#6160)
PR #6156 is failing a CI test because adding one additional int
parameter in the ImplicitPushXPSubOrbits() routine pushes the kernel
argument size beyond the 2048 byte limit.

This PR is an attempt to cleanup the ImplicitPushXPSubOrbits() function
and to hopefully minimize the size of the kernel argument as much as
possible.

@WeiqunZhang @ax3l @atmyers Any ideas for reducing this kernel argument
size?