MacroscopicEvolveE: pass macroscopic_solver_algo to MacroscopicEvolveE as a parameter instead of using WarpX::macroscopic_solver_algo (#5770)
With this PR `macroscopic_solver_algo` is passed to `MacroscopicEvolveE`
as an argument, instead of directly accessing the static variable of the
WarpX class `WarpX::macroscopic_solver_algo` .

Note that after this change `macroscopic_solver_algo` does not need to
be static. Therefore, it is renamed `m_macroscopic_solver_algo` and it
is changed into a private member variable of the WarpX class.