Docs: fix link to input files  (#6204)
Changes the links of some input files in the documentation. 
The examples in the documentation are linked to specific CI tests. 
The inputs of some of these CI tests are scattered in two separate
files: a base file, and an auxiliary one with additional options. See
for example
`../Examples/Physics_applications/laser_acceleration/`.
In the docs we now link the base files.